Break down read- methods into modules and test
